separated questionnaire from the app code
